requirements.txt
aiohttp==3.9.3
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pyarrow==15.0.0